util: improve error message for missing path_copy src
When requesting to copy a file/directory via the `path_copy` utility
call, the logic assumes its a directory if a file type is not detected.
The error in the case where `src` does not exist may not be clear to
users if the intended source was a file (i.e. the error message will
indicate the directory does not exist instead of a file). This commit
adjusts the error case to report a specific error message if the
provided source path does not exist.

+++ releng_tool/util/io.py
@@ -471,6 +471,7 @@ def path_copy(src, dst, quiet=False, critical=True):
         SystemExit: if the copy operation fails with ``critical=True``
     """
     success = False
+    errmsg = None
 
     try:
         if os.path.isfile(src):
@@ -478,13 +479,18 @@ def path_copy(src, dst, quiet=False, critical=True):
             if parent_dir and not os.path.isdir(parent_dir):
                 ensure_dir_exists(parent_dir, quiet=quiet)
             copy2(src, dst)
-        else:
+            success = True
+        elif os.path.exists(src):
             copy_tree(src, dst)
-        success = True
+            success = True
+        else:
+            errmsg = 'source does not exist: {}'.format(src)
     except (DistutilsFileError, IOError) as e:
-        if not quiet:
-            err('unable to copy source contents to target location')
-            err('    {}'.format(e))
+        errmsg = str(e)
+
+    if not quiet and errmsg:
+        err('unable to copy source contents to target location\n'
+            '    {}'.format(errmsg))
 
     if not success and critical:
         sys.exit(-1)
